Initial legend picker
last value
